pyproject.toml
[project]
name = "my-llm-writing-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "dropbox>=12.0.2",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "pandas>=2.3.3",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "pgvector>=0.4.2",
    "psycopg>=3.3.2",
    "sentence-transformers>=5.2.0",
]
